scripts/explain_verbs/requirements.txt
openai
python-dotenv
gradio
fastapi
uvicorn
jinja2
